+++ src/test/java/jp/pay/model/TermTest.java
@@ -21,10 +21,9 @@
  * THE SOFTWARE.
  *
  */
-package jp.pay;
+package jp.pay.model;
 
 
-import jp.pay.model.*;
 import jp.pay.net.APIResource;
 
 import org.junit.Test;
@@ -35,9 +34,10 @@ import java.util.List;
 import java.util.Map;
 
 import jp.pay.exception.PayjpException;
+import jp.pay.BasePayjpTest;
 import static org.junit.Assert.assertEquals;
 
-public class PayjpTermTest extends BasePayjpTest {
+public class TermTest extends BasePayjpTest {
 	@Test
 	public void testDeserialize() throws PayjpException, IOException {
 		String json = resource("model/term.json");
